Update closer to 0.44
Ok so this isn's really the 0.44 update, but its headed in the right
direction. I have added Elder Tree support, and I have started a curly
root support with root burning. But using the curly root part is ill
advised unless you know what you are doing, and modify the code to
work...

Now elder's are slightly different. Because elder trees take so damn
long to respawn, I've made a pattern following method, this means that
you are able to give it a set of locations, and the bot will travel
through each one in turn, chopping down the elder trees as it goes.
It'll bank whenever its inventory is filled too.

To enter the pattern into the bot, follow this method:
In the pattern entry textbox, enter a set of numbers separated by
comma's, DO NOT ENTER ANY LETTERS OR OTHER PUNCTUATION otherwise the bot
will not be able to read the pattern you have entered. the numbers will
refer to locations that the bot may go to, so far the supported
locations are: